Fix the unit tests test=develop
